MainView를 만들어 화면 transition 되도록 변경
- App은 MainView를 보여주고 그 안에서 화면 전환
