+++ Mulimee/Mulimee/MainView.swift
@@ -1,0 +1,36 @@
+//
+//  MainView.swift
+//  Mulimee
+//
+//  Created by Kyeongmo Yang on 7/9/24.
+//
+
+import SwiftUI
+
+struct MainView: View {
+    @State private var isLaunching = true
+    private let moveToTimer = Timer.publish(every: 2,on: .main, in: .common).autoconnect().first()
+    
+    var body: some View {
+        if isLaunching {
+            SplashView()
+                .transition(.opacity)
+                .onReceive(moveToTimer) { _ in
+                    withAnimation {
+                        isLaunching.toggle()
+                    }
+                }
+        } else {
+            let repository = RepositoryImpl()
+            let drink = repository.fetchDrink()
+            DrinkView()
+                .environment(DrinkViewModel(drink: drink,
+                                            repository: repository))
+                .transition(.opacity)
+        }
+    }
+}
+
+#Preview {
+    MainView()
+}
